+++ libtiledbsoma/src/soma/column_buffer.h
@@ -292,7 +292,7 @@ class ColumnBuffer {
     /**
      * @brief Resize the internal buffers to the given size.
      */
-    void resize(size_t size, size_t num_cells, bool preserve_data = false);
+    void resize(uint64_t size, uint64_t num_cells, bool preserve_data = false);
 
    protected:
     size_t num_cells_;
